requirements.txt
unipath
Django==1.3.1
South==0.7.3
django-social-auth==0.6.6
django-bootstrap-form
markdown==2.1.1
psycopg2==2.4.5
django-heroku-postgresify==0.2
gunicorn=0.14.2
-e git://github.com/bartTC/django-wakawaka.git#egg=wakawaka
